corrected the code according to the comments
