+++ src/components/chatWindow/messages/mes-modal/JoinChatModal.jsx
@@ -31,7 +31,7 @@ export default function JoinChatModal() {
                 <div className="modal-chat__header">
                     <h3 className="text-inter-18-600">{t('chat-context-menu.joinToChat')}</h3>
                     <MdClose
-                        className="modal-chat__close close-img"
+                        className="modal-chat__close close-img cursor-pointer"
                         size={24}
                         onClick={onCloseClick} />
                 </div>
