--- src/extra/config/theme.js
@@ -1,10 +1,0 @@
-export const theme = {
-    dark: {
-        title: 'dark',
-        value: 1,
-    },
-    light: {
-        title: 'light',
-        value: 0,
-    }
-}
