+++ src/styles/chat-window/dialogs/dialogs.scss
@@ -7,7 +7,7 @@
 }
 
 .dialogs {
-    padding-left: 10px;
+    padding: 0 0 0 10px;
 
     &__logo {
         // .dialog__logo
