+++ src/styles/static-classes.scss
@@ -226,7 +226,6 @@
 }
 
 .close-img {
-    cursor: pointer;
     stroke-width: 0.1px;
     stroke: var(--color-white-black);
 
@@ -237,7 +236,6 @@
 }
 
 .close-registration-img {
-    cursor: pointer;
     stroke-width: 0.1px;
     stroke: var(--color-violet-black);
     font-size: 24px;
